+++ lib/src/classes/timestamp.dart
@@ -5,10 +5,24 @@ import 'package:bson/src/utils/statics.dart';
 class Timestamp {
   Timestamp([int? _seconds, int? _increment])
       : seconds = _seconds ??= Statics.secondsSinceEpoch,
-        increment = _increment ?? Statics.nextIncrement;
+        increment = _nextIncrement(_seconds, _increment);
 
   final int seconds;
   final int increment;
+
+  static int _referenceSecondsSinceEpoch = 0;
+  static int _increment = 0;
+  static int _nextIncrement(int seconds, int? incParm) {
+    if (incParm != null) {
+      _referenceSecondsSinceEpoch = seconds;
+      return _increment = incParm;
+    }
+    if (seconds == _referenceSecondsSinceEpoch) {
+      return ++_increment;
+    }
+    _referenceSecondsSinceEpoch = seconds;
+    return _increment = 1;
+  }
 
   @override
   int get hashCode => '$seconds-$increment'.hashCode;